requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via
    #   django
    #   scout-apm
certifi==2021.10.8
    # via
    #   requests
    #   urllib3
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via pip-tools
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==36.0.1
    # via
    #   pyopenssl
    #   urllib3
dj-database-url==0.5.0
    # via django-heroku
django==4.0
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   django-filter
    #   django-heroku
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==3.10.1
    # via -r requirements.in
django-crum==0.7.9
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-heroku==0.3.1
    # via -r requirements.in
django-json-widget==1.1.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.0.0
    # via -r requirements.in
drf-yasg==1.20.0
    # via -r requirements.in
future==0.18.2
    # via django-json-widget
idna==3.3
    # via
    #   requests
    #   urllib3
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
markupsafe==2.0.1
    # via jinja2
packaging==21.3
    # via drf-yasg
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in
psutil==5.9.0
    # via scout-apm
psycopg2==2.9.3
    # via django-heroku
psycopg2-binary==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.3.0
    # via djangorestframework-simplejwt
pyopenssl==21.0.0
    # via urllib3
pyparsing==3.0.6
    # via packaging
pytz==2021.3
    # via djangorestframework
requests==2.26.0
    # via
    #   -r requirements.in
    #   coreapi
ruamel.yaml==0.17.19
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
scout-apm==2.23.5
    # via -r requirements.in
six==1.16.0
    # via pyopenssl
sqlparse==0.4.2
    # via django
tomli==2.0.0
    # via pep517
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3[secure]==1.26.7
    # via
    #   requests
    #   scout-apm
wheel==0.37.1
    # via pip-tools
whitenoise==5.3.0
    # via django-heroku
wrapt==1.13.3
    # via scout-apm

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
